displayMsgByChannel - fixes trashbin errors and admin-rights
Vue - data():
displayMsgByChannel : currentAccountChannel[] to store logged in account admin-information


Vue - methods:
displayMsgByChannel : added getAccountChannel() to get information if the inlogged user is admin or not, to activate trash bin.
displayMsgByChannel : async-removeMessage() is fixed, so the scroll wont get to the bottom when deleting a message
displayMsgByChannel : removeMessageIcon() gives you the ability, as an admin, to delete every message in your own channel.
displayMsgByChannel : removeMessageIcon() is recoded and now you can delete your own messages in public channels ;)