ChannelMessages displays channel options now.
JavaSpring - Service:
ChannelService : made some changes to get @PutMapping(updating) channels to work


Vue - template:
displayMsgByChannel : added divs to hold the new channel-settings box, that includes a popup/strech of input field and two buttons.
displayMsgByChannel : the new channel option has a "gear"-icon and is placed on the top of the page, over the channel name.


Vue - data():
displayMsgByChannel : added currentChannelName:'' to display the current channels name
displayMsgByChannel : added newChannelName:'' to temporary store the new channel name
displayMsgByChannel : added showChangeNameAndLeaveChannel, a variable for visibility to show all setting, both to change channel name and/or leave channel at the same time
displayMsgByChannel : added showLeaveChannelSettings, a variable for visibility to only show leave-channel-button on click
displayMsgByChannel : added showTextFieldToChangeChannelName, a variable for visibility text field and text field button
displayMsgByChannel : added showButtonToLeaveChannel, a variable for visibility for the leave-channel-button


Vue - methods:
displayMsgByChannel : added displayChannelSettings() to check if the logged in user is admin or not. If admin, then both settings for channel name change and leaving channel will be dislayed. If not admin, only leaving the channel will be possible.
displayMsgByChannel : added changeChannelName() that will make it possible to change channel name, if the changer is an admin of the channel.


CSS:
displayChannel : fixed issues with scrollContainer for not fitting the whole parent-screen
displayMsgByChannel : added CSS to the new section on the top, input field to change channel name, button to change name (reacts with the input field) and added a leave-button to connect it to async-leaveChannel() that some of the devs added.
displayMsgByChannel : changed the page container to have more rows in its grid.
messageBox : fixed messageBoxContainer not responding to 600/1200 screen width
messageBox : changed message box from 100px to 80px for a slimmer look