Small fixes in createChannel and friendListForCreateChannel.css
$store:
added "resetForChannelCreationMyFriendsThatWillBeAddedToTheNewChannel()" to reset the array

- - - - - -

Vue - template:
addFriendAdded : disabled the "reset list"-button, for now?

- - - - - -

Vue - methods:
addFriendAdded : fixed bug where the lists never resets if someone leaves "createChannel" without creating a channel.

- - - - - -

CSS:
friendListForCreateChannel : the nickname is now fixed and is placed in the middle of the name bar, so it doesnt get stuck under the avatar.